requirements.txt
neo4j
python-dotenv
torch
transformers
accelerate
streamlit